requirements.txt
mistralai==1.12.3
python-dotenv
pdf2image
Pillow
fastapi
uvicorn
python-multipart
